requirements.txt
python-dotenv
mysql-connector-python
fastapi
uvicorn
